Implementing training to neural network

+++ Source/Neural_Network_ART/net.cpp
@@ -118,7 +118,7 @@ void Net::forwardPropagation(const std::vector<double>& inputs)
 		layers.at(0).at(i).setOutput(inputs.at(i));
 	}
 
-	for (unsigned i = 1; i < layers.size(); ++i) {
+	for (size_t i = 1; i < layers.size(); ++i) {
 		Layer& prevLayer = layers.at(i - 1);
 		for (size_t j = 0; j < layers.at(i).size() - 1; ++j) {
 			layers.at(i).at(j).feedForward(prevLayer);
